### 2024-09-26 #7289 @editorialbot
[PRE REVIEW]: fortran-src: Fortran static analysis infrastructure
**Submitting author:** @dorchard (<a href="http://orcid.org/0000-0002-7058-7842">Dominic Orchard</a>)
**Repository:** https://github.com/camfort/fortran-src/
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.16.2
