### 2024-09-26 #7289 @danielskatz
👋 @dorchard - Thanks for your submission.

1. Note that your paper does not compile. Please follow the [example paper](https://joss.readthedocs.io/en/latest/example_paper.html) and note that you can click on the error above to find out more about it. In addition, you could work on the possibly missing DOIs that editorialbot suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` after making changes to the .md file or when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.
2. Your paper is about 4.5x longer than the maximum recommended for a JOSS paper (250-1000 words). You will have to substantially cut this before we can start to review it, or perhaps consider if JOSS is the right venue for your work. If you decide to cut it, you can use the command `@editorialbot check repository` to run a number of checks, including the word count for the paper.

I'm going to mark this as paused for now.
