### 2024-09-27 #7289 @dorchard
Thanks @danielskatz - I'll have a look at cutting. Does code snippets count towards the word count?
