### 2024-09-27 #7289 @danielskatz
I think they do get counted, which is why the 250-1000 is a rough expectation
