### 2024-10-22 a4fb7fe
shortened joss and fixing dois
