+++ joss/paper.md
@@ -51,25 +51,24 @@ bibliography: paper.bib
 # Summary
 fortran-src is an open source Haskell library and command-line application
 for the lexing, parsing,
-and static analysis of Fortran source code. It provides an essential
-front-end interface to build other Fortran language tools, e.g., tools for
+and static analysis of Fortran source code. It provides an
+ interface to build other Fortran language tools, e.g., for
 static analysis, automated refactoring, verification, and compilation.
-The tool provides multiple parsers which support Fortran source code conforming to the
-FORTRAN 66, FORTRAN 77, Fortran 90 and Fortran 95 standards, as well
-as some legacy extensions and partial Fortran 2003 support. The
-parsers generate a shared Abstract
+The library provides multiple parsers which support Fortran code conforming to the
+FORTRAN 66, FORTRAN 77, Fortran 90, Fortran 95 standards,
+some legacy extensions, and partial Fortran 2003 support.
+The parsers generate a shared Abstract
 Syntax Tree representation (AST), over which a variety
-of core static analyses are defined to facilitate the development of
+of core static analyses are defined to facilitate development of
 analysis and language tools.
-The library has been deployed in a number of
-projects in both academia and industry to help build further language tools
-for manipulating Fortran.
+The library has been deployed in several
+language tool projects in academia and industry.
 
 # Statement of need
 
 As one of the oldest surviving programming languages [@backus1978history], Fortran
-is used in a vast amount of software still in deployment. Fortran is not only a mainstay
-of legacy software, but is also still used to write new software,
+underpins a vast amount of software still in deployment. Fortran is not only a mainstay
+of legacy software, but is also used to write new software,
 particularly in the sciences. Given the importance of numerical
 models in science, verifying the correctness of such models is
 critical for scientific integrity and progress. However, doing so is
@@ -77,7 +76,7 @@ difficult, even more so than for traditional software; for
 computational models, the expected program behaviour is often unknown,
 uncertainty is the rule, and approximations are pervasive. Despite
 decades of progress in program verification within computer science,
-few formal verification techniques are currently applied in scientific software. To
+few formal verification techniques are applied in scientific software. To
 facilitate a step-change in the effectiveness of verification for
 computational science, a subset of the authors of this paper
 developed a suite of verification and static
@@ -93,31 +92,30 @@ decisions [@mendez2014climate]. In 2024, Fortran re-entered the Top 10 programmi
 the [TIOBE Index](https://www.tiobe.com/tiobe-index/), showing its enduring popularity.
 
 The continued use of Fortran, particualarly in
-scientific contexts, was the catalyst for the fortran-src software package.
+scientific contexts, was the catalyst for this software package.
 
-One of the challenges in writing language tools for Fortran is its long
+A challenge in writing language tools for Fortran is its long
 history. There have been several major language standards (FORTRAN
-I-IV, FORTRAN 66 and 77, Fortran 90, 95, 2003, 2008, 2018 and more) or
+I-IV, FORTRAN 66 and 77, Fortran 90, 95, 2003, 2008, etc.) or
 _restandardisations_. Newer standards often deprecate features
-which were known to be a ready source of error, or were difficult to
+which were known to be a ready source of errors, or were difficult to
 specify or understand. However, compilers often support an amalgam of features across
[…]
