+++ joss/paper.bib
@@ -28,24 +28,6 @@
   bibsource = {dblp computer science bibliography, http://dblp.org}
 }
 
-@article{DBLP:journals/jocs/OrchardRO15,
-  author    = {Orchard, Dominic and
-               Rice, Andrew and
-               Oshmyan, Oleg},
-  title     = {Evolving Fortran types with inferred units-of-measure},
-  journal   = {J. Comput. Science},
-  volume    = {9},
-  pages     = {156--162},
-  year      = {2015},
-  url       = {http://dx.doi.org/10.1016/j.jocs.2015.04.018},
-  doi       = {10.1016/j.jocs.2015.04.018},
-  timestamp = {Mon, 13 Jul 2015 11:12:24 +0200},
-  biburl    = {http://dblp.dagstuhl.de/rec/bib/journals/jocs/OrchardRO15},
-  bibsource = {dblp computer science bibliography, http://dblp.org}
-}
-
-
-
 @article{DBLP:journals/cse/ContrastinRDO16,
   author    = {Mistral Contrastin and
                Andrew C. Rice and
@@ -115,14 +97,21 @@
 }
 
 @article{orchard2017verifying,
-  title={Verifying spatial properties of array computations},
-  author={Orchard, Dominic and Contrastin, Mistral and Danish, Matthew and Rice, Andrew},
-  journal={Proceedings of the ACM on Programming Languages},
-  volume={1},
-  number={OOPSLA},
-  pages={75},
-  year={2017},
-  publisher={ACM}
+  author       = {Dominic A. Orchard and
+                  Mistral Contrastin and
+                  Matthew Danish and
+                  Andrew C. Rice},
+  title        = {Verifying spatial properties of array computations},
+  journal      = {Proc. {ACM} Program. Lang.},
+  volume       = {1},
+  number       = {{OOPSLA}},
+  pages        = {75:1--75:30},
+  year         = {2017},
+  url          = {https://doi.org/10.1145/3133899},
+  doi          = {10.1145/3133899},
+  timestamp    = {Sat, 30 Sep 2023 10:23:24 +0200},
+  biburl       = {https://dblp.org/rec/journals/pacmpl/OrchardCDR17.bib},
+  bibsource    = {dblp computer science bibliography, https://dblp.org}
 }
 
 @inproceedings{contrastin2016lightning,
@@ -170,27 +159,39 @@
 
 
 @article{backus1978history,
-  title={The history of {F}ortran {I}, {II}, and {III}},
-  author={Backus, John},
-  journal={ACM Sigplan Notices},
-  volume={13},
-  number={8},
-  pages={165--180},
-  year={1978},
-  publisher={ACM New York, NY, USA}
+author = {Backus, John},
+title = {The history of {FORTRAN I, II, and III}},
+year = {1978},
+issue_date = {August 1978},
+publisher = {Association for Computing Machinery},
+address = {New York, NY, USA},
+volume = {13},
+number = {8},
+issn = {0362-1340},
+url = {https://doi.org/10.1145/960118.808380},
+doi = {10.1145/960118.808380},
[…]
