+++ joss/paper.md
@@ -49,40 +49,22 @@ bibliography: paper.bib
 
 
 # Summary
-fortran-src is an open source Haskell library and command-line application
+fortran-src is an open-source Haskell library and command-line application
 for the lexing, parsing,
 and static analysis of Fortran source code. It provides an
  interface to build other Fortran language tools, e.g., for
 static analysis, automated refactoring, verification, and compilation.
-The library provides multiple parsers which support Fortran code conforming to the
-FORTRAN 66, FORTRAN 77, Fortran 90, Fortran 95 standards,
-some legacy extensions, and partial Fortran 2003 support.
-The parsers generate a shared Abstract
-Syntax Tree representation (AST), over which a variety
-of core static analyses are defined to facilitate development of
-analysis and language tools.
+The library supports FORTRAN 66, FORTRAN 77, Fortran 90, Fortran 95,
+some legacy extensions, and partially Fortran 2003, with
+a shared Abstract Syntax Tree representation.
 The library has been deployed in several
 language tool projects in academia and industry.
 
 # Statement of need
 
 As one of the oldest surviving programming languages [@backus1978history], Fortran
-underpins a vast amount of software still in deployment. Fortran is not only a mainstay
-of legacy software, but is also used to write new software,
-particularly in the sciences. Given the importance of numerical
-models in science, verifying the correctness of such models is
-critical for scientific integrity and progress. However, doing so is
-difficult, even more so than for traditional software; for
-computational models, the expected program behaviour is often unknown,
-uncertainty is the rule, and approximations are pervasive. Despite
-decades of progress in program verification within computer science,
-few formal verification techniques are applied in scientific software. To
-facilitate a step-change in the effectiveness of verification for
-computational science, a subset of the authors of this paper
-developed a suite of verification and static
-analysis tools named CamFort to explore lightweight verification methods
-(requiring little or no programmer effort), targeted at
-scientific programming [@contrastin2016lightning]. We chose Fortran as it remains a popular
+underpins a vast amount of software; Fortran is not only a mainstay
+of legacy software, but is also used to write new software. Fortran remains a popular
 language in the international scientific community; @vanderbauwhede2022making
 reports data from 2016 on the UK's \`\`Archer'' supercomputer, showing the
 vast majority of use being Fortran code. Fortran is
@@ -90,19 +72,18 @@ particularly notable for its prevalence in earth sciences, e.g., for
 implementing global climate models that then inform international policy
 decisions [@mendez2014climate]. In 2024, Fortran re-entered the Top 10 programming languages in
 the [TIOBE Index](https://www.tiobe.com/tiobe-index/), showing its enduring popularity.
-
-The continued use of Fortran, particualarly in
+The continued use of Fortran, particularly in
 scientific contexts, was the catalyst for this software package.
 
 A challenge in writing language tools for Fortran is its long
 history. There have been several major language standards (FORTRAN
 I-IV, FORTRAN 66 and 77, Fortran 90, 95, 2003, 2008, etc.) or
 _restandardisations_. Newer standards often deprecate features
-which were known to be a ready source of errors, or were difficult to
+which are known to be a ready source of errors, or were difficult to
 specify or understand. However, compilers often support an amalgam of features across
-language standards, including deprecated features (@urmaetal2014).
+standards (@urmaetal2014).
 This enables developers to keep using deprecated features, or mix
-a variety of language standards.
+language standards.
 This complicates the task of developing new tools for manipulating Fortran
[…]
