+++ joss/paper.md
@@ -2,12 +2,12 @@
 title: 'fortran-src: Fortran static analysis infrastructure'
 tags:
   - Fortran
-
-# TODO should attempt to add all relevant authors
+  - static analysis
+  - tooling
 authors:
 - name: Mistral Contrastin
   orcid: 0000-0002-5409-7122
-  affiliation: 1,2
+  affiliation: "1, 2"
 - name: Raoul Hidalgo Charman
   orcid: 0000-0002-8401-7672
   affiliation: 3
@@ -19,57 +19,53 @@ authors:
   affiliation: 5
 - name: Dominic Orchard
   orcid: 0000-0002-7058-7842
-  affiliation: 5,6
+  affiliation: "5, 6"
 - name: Andrew Rice
   orcid: 0000-0002-4677-8032
-  affiliation: 1,7
+  affiliation: "1, 7"
 - name: Jason Xu
   orcid: 0000-0003-3310-0756
   affiliation: 3
 
 affiliations:
-- name: Department of Computer Science and Technology, University of Cambridge
+- name: Department of Computer Science and Technology, University of Cambridge, UK
   index: 1
-- name : Meta
+- name : Meta, US
   index: 2
-- name : Bloomberg
+- name : Bloomberg, US
   index: 3
-- name : Utrecht University
+- name : Utrecht University, Netherlands
   index: 4
-- name: School of Computing, University of Kent
+- name: School of Computing, University of Kent, UK
   index: 5
-- name: Institute of Computing for Climate Science, University of Cambridge
+- name: Institute of Computing for Climate Science, University of Cambridge, UK
   index: 6
-- name : GitHub
+- name : GitHub, US
   index: 7
 
 date: TODO
 bibliography: paper.bib
 ---
 
-
 # Summary
+
 fortran-src is an open-source Haskell library and command-line application
 for the lexing, parsing,
 and static analysis of Fortran source code. It provides an
- interface to build other Fortran language tools, e.g., for
+ interface to build other tools, e.g., for
 static analysis, automated refactoring, verification, and compilation.
 The library supports FORTRAN 66, FORTRAN 77, Fortran 90, Fortran 95,
 some legacy extensions, and partially Fortran 2003, with
 a shared Abstract Syntax Tree representation.
-The library has been deployed in several
-language tool projects in academia and industry.
+The library has been deployed in several language tool projects in academia and industry.
 
 # Statement of need
 
-As one of the oldest surviving programming languages [@backus1978history], Fortran
-underpins a vast amount of software; Fortran is not only a mainstay
-of legacy software, but is also used to write new software. Fortran remains a popular
-language in the international scientific community; @vanderbauwhede2022making
-reports data from 2016 on the UK's \`\`Archer'' supercomputer, showing the
+As one of the oldest surviving programming languages [@backus1978history], Fortran underpins a vast amount of software; Fortran is not only a mainstay
+of legacy software, but is also used to write new software. Fortran remains a popular language in the international scientific community; @vanderbauwhede2022making reports data from 2016 on the UK's \`\`Archer'' supercomputer, showing the
 vast majority of use being Fortran code. Fortran is
 particularly notable for its prevalence in earth sciences, e.g., for
-implementing global climate models that then inform international policy
+implementing global climate models that inform international policy
 decisions [@mendez2014climate]. In 2024, Fortran re-entered the Top 10 programming languages in
 the [TIOBE Index](https://www.tiobe.com/tiobe-index/), showing its enduring popularity.
 The continued use of Fortran, particularly in
@@ -77,17 +73,16 @@ scientific contexts, was the catalyst for this software package.
 
 A challenge in writing language tools for Fortran is its long
[…]
