+++ joss/paper.bib
@@ -6,25 +6,10 @@
                2013, Indianapolis, IN, USA, October 27, 2013},
   pages     = {29--32},
   year      = {2013},
-  crossref  = {DBLP:conf/oopsla/2013wrt},
   url       = {http://doi.acm.org/10.1145/2541348.2541356},
   doi       = {10.1145/2541348.2541356},
   timestamp = {Mon, 08 Sep 2014 09:39:24 +0200},
   biburl    = {http://dblp.dagstuhl.de/rec/bib/conf/oopsla/OrchardR13},
-  bibsource = {dblp computer science bibliography, http://dblp.org}
-}
-
-@proceedings{DBLP:conf/oopsla/2013wrt,
-  editor    = {Emerson R. Murphy{-}Hill and
-               Max Sch{\"{a}}fer},
-  title     = {Proceedings of the 2013 {ACM} Workshop on Refactoring Tools, WRT@SPLASH
-               2013, Indianapolis, IN, USA, October 27, 2013},
-  publisher = {{ACM}},
-  year      = {2013},
-  url       = {http://dl.acm.org/citation.cfm?id=2541348},
-  isbn      = {978-1-4503-2604-9},
-  timestamp = {Mon, 08 Sep 2014 09:39:24 +0200},
-  biburl    = {http://dblp.dagstuhl.de/rec/bib/conf/oopsla/2013wrt},
   bibsource = {dblp computer science bibliography, http://dblp.org}
 }
 
@@ -69,6 +54,7 @@
                Andrew C. Rice and
                Dominic Orchard},
   title     = {Learning units-of-measure from scientific code},
+  doi = {0.1109/SE4Science.2019.00013},
   booktitle = {Proceedings of the 14th International Workshop on Software Engineering
                for Science, SE4Science@ICSE 2019, Montreal, QC, Canada, May 28, 2019},
   pages     = {43--46},
