### 2024-11-07 #7289 @dorchard
@danielskatz the build and DOIs are fixed now (apologies) and I have made the paper considerably shorter (less than half the length); would that be acceptable? It's about 5 pages and I wasn't sure about sacrificing a clear description of the library features, structure, and major uses of the library. Is this reviewable now?
