+++ joss/paper.bib
@@ -266,17 +266,6 @@ url = {https://github.com/wimvanderbauwhede/RefactorF4Acc},
 note = {Retrieved November 2023}
 }
 
-@article{vanderbauwhede2022making,
-  title={Making legacy Fortran code type safe through automated program transformation},
-  author={Vanderbauwhede, Wim},
-  journal={The Journal of Supercomputing},
-  volume={78},
-  number={2},
-  pages={2988--3028},
-  year={2022},
-  publisher={Springer}
-}
-
 @misc{danish2024incremental,
   author       = {Matthew Danish and
                   Dominic Orchard and
