### 2024-11-11 #7289 @danielskatz
@dorchard - the paper is still showing as 2200 words, which is far more than the JOSS guidelines of 250-1000. Something more needs to be removed. I suggest shortening information about the features, and pointing to something in the repo or docs.
