+++ joss/paper.md
@@ -55,8 +55,7 @@ and static analysis of Fortran source code. It provides an
  interface to build other tools, e.g., for
 static analysis, automated refactoring, verification, and compilation.
 The library supports FORTRAN 66, FORTRAN 77, Fortran 90, Fortran 95,
-some legacy extensions, and partially Fortran 2003, with
-a shared Abstract Syntax Tree representation.
+some legacy extensions, and partially Fortran 2003.
 The library has been deployed in several language tool projects in academia and industry.
 
 # Statement of need
@@ -65,7 +64,7 @@ As one of the oldest surviving programming languages [@backus1978history], Fortr
 of legacy software, but is also used to write new software. Fortran remains a popular language in the international scientific community; @vanderbauwhede2022making reports data from 2016 on the UK's \`\`Archer'' supercomputer, showing the
 vast majority of use being Fortran code. Fortran is
 particularly notable for its prevalence in earth sciences, e.g., for
-implementing global climate models that inform international policy
+implementing climate models that inform international policy
 decisions [@mendez2014climate]. In 2024, Fortran re-entered the Top 10 programming languages in
 the [TIOBE Index](https://www.tiobe.com/tiobe-index/), showing its enduring popularity.
 The continued use of Fortran, particularly in
@@ -77,8 +76,8 @@ I-IV, FORTRAN 66 and 77, Fortran 90, 95, 2003, 2008, etc.).
 Newer standards often deprecate features
 which are known to be a ready source of errors, or difficult to
 specify or understand. However, compilers often support an amalgam of features across
-standards (@urmaetal2014).
-This enables developers to keep using deprecated features and mix
+standards (@urmaetal2014),
+enabling developers to keep using deprecated features and mix
 language standards. This complicates the development of new tools for manipulating Fortran
 source code; one must tame the weight of decades of language evolution.
 
@@ -107,14 +106,21 @@ tools for refactoring Fortran [@vanderbauwhede2022making]:
 No comprehensive lexing, parsing, and analysis library was available from which to
 build new tools.
 
-# Functionality
+# Functionality in brief
 
-  * Lexing and parsing of Fortran to an expressive Abstract Syntax Tree;
-  * Various static analyses, e.g., data flow analysis;
+  * Lexing (of both fixed and free form code) and parsing of Fortran to an expressive unified Abstract Syntax Tree;
+  * Static analyses, e.g., general data flow analysis including:
+    - Reaching definitions;
+    - Def-use/use-def;
+    - Constant evaluation;
+    - Constant propagation;
+    - Live variable analysis;
+    - Induction variable analysis.
   * Type checking;
+  * Module graph analysis;
   * Pretty printing;
-  * "Reprinting", or patching sections of source code without removing secondary
-    notation such as comments;
+  * "Reprinting" (patching sections of source code without removing secondary
+    notation such as comments);
   * Exporting to JSON.
 
 fortran-src is primarily a Haskell library but it also packages a command-line
@@ -122,110 +128,10 @@ tool. By exporting parsed code to JSON, the
 parsing and analyses that fortran-src provides may be utilized by
 non-Haskell tools.
 
-The library's top-level module is `Language.Fortran`.
-
-## Lexing and parsing
-
-Static analysis of Fortran requires a choice in the lexing and parsing
-front end: either to take the approach of many compilers, allowing an amalgam of features (e.g., gfortran with
-its hand-written parser), or to
-enforce language standards at the exclusion of some code that is
-accepted by major compilers. fortran-src takes roughly the latter
-approach, though it has an extended Fortran 77 mode for supporting
-legacy extensions influenced by vendor-specific compilers popular in the past.
-
-The Fortran language has evolved through two broad syntactic forms:
-
[…]
