### 2024-12-03 #7289 @dorchard
Now 1477, and about only just over 3 pages excluding references. Would this be acceptable even if it is a little over?
