### 2024-12-03 #7289 @danielskatz
@dorchard - note that one of your references is missing a leading "1" in the doi field...

I do think the length is ok.

I'm a little unsure about the acknowledgments you've put in the middle of your license, and if this is still a valid, OSI-approved license, which I'm checking with another editor on.

Other than these things, I think the review can start.
