### 2024-12-03 #7289 @dorchard
Thanks @danielskatz . I can't see which reference you are referring to, sorry.
Re the LICENCE file- I see what you mean. I could take this out into their own file (`ACKNOWLEDGMENTS.md`).
