### 2024-12-03 #7289 @danielskatz
If you can make the license change, please do.

Re the bib entry, you have one that has doi = {0.1109/SE4Science.2019.00013} that should be {10.1109/SE4Science.2019.00013}
