+++ joss/paper.bib
@@ -54,7 +54,7 @@
                Andrew C. Rice and
                Dominic Orchard},
   title     = {Learning units-of-measure from scientific code},
-  doi = {0.1109/SE4Science.2019.00013},
+  doi = {10.1109/SE4Science.2019.00013},
   booktitle = {Proceedings of the 14th International Workshop on Software Engineering
                for Science, SE4Science@ICSE 2019, Montreal, QC, Canada, May 28, 2019},
   pages     = {43--46},
