+++ ACKNOWLEDGMENTS.md
@@ -1,0 +1,17 @@
+Additional thanks for contributions from:
+
+* Anthony Burzillo
+* Azeem Bande-Ali
+* Raoul Hidalgo Charman
+* Harry Clarke
+* Bradley Hardy
+* Aiden Jeffrey
+* Lukasz Kolodziejczyk
+* Vilem-Benjamin Liepelt
+* Ben Moon
+* Ben Orchard
+* Eric Seidel
+* Poppy Singleton-Hoare
+* TravelTissues
+* Jason Xu
+* Vaibhav Yenamandra
