### 2024-12-04 #7289 @danielskatz
Our bot checks everything in the bib file, not just what is used...

But thanks
