### 2024-12-05 #7289 @danielskatz
👋 @filippi, @fzhao70, @jabirali - Would one or more of you be willing to review this submission for JOSS?
