### 2024-12-05 #7289 @danielskatz
👋 @rouson - would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Or might you have some suggestions for possible reviewers?
