### 2024-12-05 #7289 @filippi
Hi,
I can review the paper and try the binary, but I don't know anything about
Haskell so will be very hard to co code review
regards
********************************************
    Jean-Baptiste Filippi
    Chargé de Recherches
    SPE - UMR 6134 CNRS
    Université de Corse
    Quartier grossetti
    BP 52, 20250 Corte
    Tel. +33 495 450 158
******************************************

Le jeu. 5 déc. 2024 à 17:33, Daniel S. Katz ***@***.***> a
écrit :

> 👋 @filippi <https://github.com/filippi>, @fzhao70
> <https://github.com/fzhao70>, @jabirali <https://github.com/jabirali> -
> Would one or more of you be willing to review this submission for JOSS?
>
