### 2024-12-06 #7289 @tclune
> 👋 @tclune - Would you be willing to review this submission for JOSS?

@danielskatz - Yes, presuming that I can defer the review until early January.   Burning lots of use-or-lose and that will be a quiet window.  I can sort of pretend that this is not work.  :-)
