### 2024-12-13 #7571 @filippi
# Review of *fortran-src*

**Summary**

The manuscript presents *fortran-src*, a Haskell-based library and command-line tool for analyzing, parsing, and lexing Fortran code across multiple standards. The paper is concise, well-structured, and clearly describes the motivation and capabilities of the tool. It explains the historical and current significance of Fortran, highlighting the persistent need to modernize legacy code to meet evolving standards and take advantage of newer compilers and analysis tools.

The provided references and background are appropriate. The paper also contrasts *fortran-src* with existing (often commercial or closed-source) tools, thus underscoring its place in the ecosystem of Fortran language processing tools.

**Comments on the Paper**

The paper is clear and mostly self-contained. It effectively justifies the tool’s relevance by underlining the complexity of evolving Fortran standards and the challenges of handling legacy code. The work is presented with sufficient detail. I have no major comments or suggested corrections regarding the manuscript’s content and style. It is concise, informative, and presents a well-defined need.
Tiny minor corrections :
   - In the “Related software” section, correct the spelling of “General commerical static analysis tools” to “General commercial static analysis tools.”
   - Consider consistent hyphenation for “open source” as “open-source” where it is used adjectivally (e.g., “an open-source Haskell library”).

**Tool Usage and Practicality**

Given my lack of Haskell expertise, I tested the provided binary form of *fortran-src* on a large set of Fortran 90 source files originating from a French weather forecast model. This is a non-trivial test due to the large codebase and various macro expansions employed for parallelization and architectural differences.

- **Lexing and Parsing:**
  In general, the lexing and parsing worked correctly across most of the source files. One caveat observed is that macro expansions sometimes lead to syntax that *fortran-src* cannot handle directly. Once such files were pre-processed (resulting in standard conformant Fortran 90 code), the tool parsed them mostly without issue. On some files, parsing stopped at a construct involving `TDoubleColon` tokens, for instance at `(799:63)-(799:64)`... and basically this is because i had very badly made strides in my arrays like `TARDIS(:,:,:,1::22)`, that did compile well.. but not standard and ambiguous. Thank you fortran-src.

- **Pretty Printing and Analysis:**
  The pretty-printing functionality worked as expected, consistently producing reformatted code.
  The analysis functions are promising. In particular, analyzing selected code segments to identify modernization or refactoring opportunities was straightforward and provides a helpful starting point for developers maintaining or migrating legacy Fortran code.

- **Dot files:**
  The generated dot files looks correct, but won't compile on my version of graphviz. I could not try.

**Conclusion**

*fortran-src* appears to be a mature, well-documented codebase that meets a real and persistent need in the Fortran community. Its ability to serve as both a standalone tool and as a library component for building other tools is valuable. The tool’s overall performance and capabilities are commendable. It should be a useful resource for practitioners and researchers dealing with legacy Fortran code modernization.

I recommend publication.
