### 2024-12-18 #7571 @fzhao70
# Review of fortran-src

## Manuscript Evaluation

The manuscript clearly articulates the statement of need and thoroughly describes the library's development process. The content meets the JOSS publication standards, and no further modifications to the manuscript are necessary.

## Code Evaluation

I conducted testing using both the prebuilt binary (version 0.12.0) and a local build using cabal (version 0.16.3) on Ubuntu 20.04, with Fortran code from the Weather Research and Forecasting (WRF) model physics modules and the provided test cases in the test-data folder. Both versions successfully generated the expected DOT files, produced correct reprints, and generated accurate submodule relationship graphs. I also tested the fortran-src library directly through its Haskell interface using a simple Fortran code example, which produced the expected results.

## Recommendations

I recommend that the authors provide prebuilt binaries for versions after 0.12.0 for mainstream operating systems to facilitate usage for users without cabal installed, include example use cases demonstrating CLI command usage, add example use cases showing how to use the software as a Haskell library, and enhance error reporting when parsing or lexing fails, particularly for cases involving module files.

## Summary

fortran-src is a valuable and practical library for analyzing Fortran code. I strongly recommend it for publication in JOSS.
