### 2024-12-18 #7571 @danielskatz
Thanks @fzhao70 - JOSS does not require a formal review, just that the checklist is completed, but your comments are helpful, along with your recommendations.

@dorchard - please consider these recommendations.
