### 2025-01-07 #7571 @dorchard
> Thanks @fzhao70 - JOSS does not require a formal review, just that the checklist is completed, but your comments are helpful, along with your recommendations.
>
> @dorchard - please consider these recommendations.

Yes these are helpful - I will make some edits + changes this week.
