### 2025-01-14 #7571 @dorchard
> Thanks [@filippi](https://github.com/filippi).
>
> [@dorchard](https://github.com/dorchard) - Please consider if you want to make any changes, even though [@filippi](https://github.com/filippi) has signed off on publication without saying that any are needed.

Thanks @filippi. I have fixed the typos you suggested. Re the problem with the dot file- if you could raise an issue that would be great!
