### 2025-01-14 #7571 @dorchard
> Thanks [@fzhao70](https://github.com/fzhao70) - JOSS does not require a formal review, just that the checklist is completed, but your comments are helpful, along with your recommendations.
>
> [@dorchard](https://github.com/dorchard) - please consider these recommendations.

Re the recommendations:

> I recommend that the authors provide prebuilt binaries for versions after 0.12.0 for mainstream operating systems to facilitate > usage for users without cabal installed,

We are looking into it.

> include example use cases demonstrating CLI command usage, add example use
> cases showing how to use the software as a Haskell library,

We have this in the Wiki. We say:
> Functionality and example usage of the tool and library is described in detail on the [fortran-src wiki](https://github.com/camfort/fortran-src/wiki/).
> A demonstration of fortran-src for static analysis is provided
> by a small demo tool which detects if an allocatable array is used
> before it has been allocated.\footnote{\url{https://github.com/camfort/allocate-analysis-example}}

But we could add a little to the paper if the editors were okay with it?

> and enhance error reporting when parsing or lexing fails, particularly for cases involving module files.

I've raised this an issue which hopefully we can address at some point. Does it need address now for this paper?
