### 2025-01-14 #7571 @danielskatz
> > include example use cases demonstrating CLI command usage, add example use
> > cases showing how to use the software as a Haskell library,
>
> We have this in the Wiki. We say:
>
> > Functionality and example usage of the tool and library is described in detail on the [fortran-src wiki](https://github.com/camfort/fortran-src/wiki/).
> > A demonstration of fortran-src for static analysis is provided
> > by a small demo tool which detects if an allocatable array is used
> > before it has been allocated.\footnote{\url{[https://github.com/camfort/allocate-analysis-example}}](https://github.com/camfort/allocate-analysis-example%7D%7D)
>
> But we could add a little to the paper if the editors were okay with it?

I think having it in the wiki is sufficient, but I wonder if @fzhao70 thinks anything in the paper that points to it should be changed?

> > and enhance error reporting when parsing or lexing fails, particularly for cases involving module files.
> I've raised this an issue which hopefully we can address at some point. Does it need address now for this paper?

I think this is also a question for @fzhao70 - Do you think addressing this is a requirement for publication, or just a suggestion to improve the software that can be as future work?
