### 2025-01-20 #7571 @tclune
@danielskatz I'm feeling stupid - where do I find the `paper.md` file.   It is not in the downloaded source.  Other reviewers (and even the bot!) found it ...
