### 2025-01-20 #7571 @tclune
# Review by T. Clune

## Manuscript evaluation

The paper was well written and provided sufficient references.   No modifications requested.

## Code evaluation

Due to time limitations, I only performed perfunctory tests - working through some of the documented examples.   I did find one minor issue related to default response, and will report that as a GitHub issue in `fortran-src`.

## Documentation

Although I checked the relevant boxes, I found that the in-repo documentation for "Statement of Need" could be improved.
 The capabilities are described at a very high level, but nothing suggests who the target audience is nor why they might need this capability beyond "Fortran users".   The paper does a nice job of this, so it should be easy to bring some of that text into the README file.

## Summary

I recommend this software for publication in JOSS.
