### 2025-01-20 #7571 @danielskatz
@dorchard - please see the comments from @tclune above, and consider responding to them.  I'm going to take them just as recommendations, and not requirements, and will now continue the JOSS process by proofreading the paper.
