### 2025-01-20 #7571 @dorchard
Merged! Thanks for these.
