### 2025-01-20 #7571 @dorchard
Thanks @tclune. I have updated the readme with more info on statement of need and features.
