+++ package.yaml
@@ -1,5 +1,5 @@
 name: fortran-src
-version: '0.16.3'
+version: '0.16.4'
 synopsis: Parsers and analyses for Fortran standards 66, 77, 90, 95 and 2003 (partial).
 description: >-
   Provides lexing, parsing, and basic analyses of Fortran code covering
