### 2025-02-07 #7571 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
